Began implementation of Modal component.
Thus far, only the props have been tested for showing and hiding the Modal component. The component still needs much work before it is considered done. Such as: testing the button under the context of being controlled by another component/element and the Modal gaining sole focus of the application.