refactor: Make GroupTable warning message optional
Change the warning message of the GroupTable to be optional in order
to assist with test brevity. This also allows a more obvious warning
message by default if flowery language is not needed/wanted.

Secondarily, update some tests for GroupTable to be more focused on
their testing scenario and remove re-testing what a more specified test
would do. Thus, moving toward a goal where queries assist with testing
almost as much as explicit asserts.