feat: Add rendering of the boundary MOV of a group
Add the display of the lowest and highest movement rating values from
the list of participants in a shared group. Compares all the
participant's movement ratings and keeps key values for display. This
helps the user quickly identify who acts as a base for the derived
speeds since the slowest member in the chased party always has one
speed point and acts as the base for all other members in a chase.