refactor: Update some packages
Update various packages and bump a few packages such as
@testing-library/react to their latest semver. @testing-library/react
received a bump in major due to the addition of the 'keyboard' method.
This update also comes with the need to rewrite a test involving the
'type' method due to a breaking change.

Also, change a few config files to better fit the updated associated
packages they configured.