feat: Attach a modal to the delete group action
Add a delete confirmation modal to GroupTable so that an accidental
click can no longer removes all stored data on a group, just like
ParticipantTable.

Also, remove the keyboard support for GroupRow which
trapped keyboard inputs because it caused a small bug when prompting
the delete modal, and the proper method of using the keyboard to
support a grid role for screen readers is not fully understood at this
time.