feat: Add classes for boundary Movement Ratings
Add styling to participants with the highest and lowest movement ratings
based on their group's collection of participants. This helps the user
visually understand which members of a group cause that group to
become slower or faster while also highlighting within the list to help
associate a name to the value.