feat: Implement the ability to add participants to groups
Add event handlers to the add button for participants so that the user
may use participants in the group to calculate their relative action
count in a chase. The participants are made available through a modal
which allows the user to select any participants that are not already a
part of that group.

Also, add a launch.json for VS Code along with VS Code's .gitignore.