refactor: Change the names of color variables
Update the names of the CSS variables to help with scalability and limit
the rigidity on what colors should be used where. Color names are
provided from "coolors.co", and used to give variables a name. The rgb
values are also extracted from the rgb functions so that they may be
used in rgba functions as well which increases the scalability.