refactor: Conform variable names to desired lexicon
Change the numerical value of a statistic to be known as "score,"
instead of, a generic "value." Value could be confused with a new
name instead of the score beforehand. Thus, values -- once they reach a
point where they need to be calculated -- turn into a score for usage.
This includes: calculating the need for styling in relation to a related
limiter.

Secondly, update any variable with "skill" or "stat" in its name to
uniformly have "statistic" instead. This is to avoid confusion between
what both of these words represent as they represent the same; meaning
they should have the same label.

Lastly, change the variables called "evt" into a fully spelled out "event"
in order to avoid possible confusion of what "evt" means.