style: Update modals to clearly separate sections
Add <hr> elments as well as using <h#> elements to clearly show --
semantically and visually -- the structure of the various modals.

Also, add various FIXMEs and READMEs to help document what needs to be
done in the code. This refers to various bugs and oversights, instead
of, brand new features.