refactor: Update lexicon for variables
Reinforce the correct use of language when manipulating data. "Adding"
something infers that it ALREADY exists and is being appended to another
object. "Removing" something infers that it ALREADY exists and is being
deducted from another object. "Creating" something infers that it DOES
NOT exist and is being made into existance. "Deleting" something infers
that something ALREADY exists and is being completely removed from
existance.