feat: Remove the option to add assigned participants
Remove the feature to assign participants to multiple groups. A
participant is a single entity and should be represented as such since
it may be involved in only one chase. If for any reason a participant
should "split" in order to be in multiple chases, a new participant
should be made to represent this. To make this feature work, a boolean
is used to represent if the participant is an orphan/free-agent.