refactor: Update the flag for an owned participant
Change the flag of an owned participant when it gets added to a group.
The participant and group objects are mutable due to their nature of
changing throughout the entire project. The groups and participants
should stay synced at all times; meaning the individuals are known to
change due to input fields. Creating a new object on all changes would
be counterintuative to the needs of this project, and would cause more
effort than is worth to keep synced otherwise.

From here on out, groups and participants will be known to be mutated
directly and will have per-line comments so that eslint ignores prop
mutations.