feat: Add an icon to show important members
Add an icon to show which members are important. Important meaning: the
participant with the lowest and the participant with the highest
movement rates. The icon is a supplement for those who may be
colorblind so that they may have a visual cue independent of colors.

Also, remove the background color for members with boundary values and
change the font color instead. This assists in emphasizing the visual
draw to important members while also color coding said members for quick
reference.