refactor: Update the style of groups' member sections
Make the the member section of groups more compact friendly. Added
symbols to label key information so that even colorblind users can
understand what is meant to be portrayed. The color coding is now an
extra emphasize instead of the only way to portray the key info.

Also, change the information to be displayed in a table instead of a
list. This grants the added benefit of having headers to help navigate
and organize the collected data. This comes accompanied with other
aria-* attributes to help bind the table with its associated error
message while barren of participants to display as members.