feat: Initially hide the distancer warning message
Hide the warning message for the distancer combobox until the user blurs
the element. This is because the initial error may make it seem like
possessing no distancer is an erroneous state instead of the abscence of
data which is a valid state for the distancer.

Also, remove associated tests that presumed that the warning message
should render initially.