feat: Connect the pursuers element to its warning
Connect the pursuer to its warning message when the message is visible.
This will help screen readers understand the purpose of the warning
message. The message will be used as a description of the current state
of the pursuer element while still occupying space -- preventing the
element from being visually void.