feat: Implement the basics of group merging
Allow the user to select another group to combine with the initiating
group of the merging process. This process deletes the other,
subservient group and moves its members to the initial group. This
will help the user quickly move all members from one group to another.

Also, simplified the event handlers shared between GroupContainer,
GroupTable, and TabbedDisplay to only use "handleGroupsChange" to update
the display instead of passing a handlers to grandchildren. The state
changes are now bubbled up to the parent component resulting in
TabbedDisplay still managing the state without the specific use-case
event handlers polluting the code space.

Finally, add some utils to help simplify the creation of groups and
participants. This should allow for a more rubust testing experience
since the creation process now has the capabilities to track ids and
type the created associated objects.