feat: Add skeletal functionality of splitting
Simply, allow the user to open and close the modal for the splitting
process. This is incremental so that the visual design is simple and
obvious for the user before implementing the main functionality
of the modal's purpose. That purpose being to shift members from an
existing group into a newly created group when dragging members and
confirming the actions in the modal.