feat: Add the switching of members between groups
This feature will allow for users to succinctly create a new group from
pre-existing groups, instead of, forcing the user to create a new group,
and then forcing the user to individually go through the participant
adding-and-removing process. The user, through this implemented feature,
simply clicks the split button and clicks another button with the
associated participant to move that desired participant to and from the
newly generated group in a toggling manner.