feat: Complete the process of splitting a group
Allow the user to split a group to ease the process of managing an
evolving chase environment. The user chooses a group to split, and then
selects all the members of the initializing group that they want to move
into a new group which they name before commiting to the splitting
action.