test: Update tests to conform to team guidelines
Use the newly defined guidelines to write more predicatble tests. This will allow developers to easily organize and write tests that are useful for them and, more importantly, the end user. Some tests as a result have been trimmed to erase information not relevant to the end user. We adhere to the principle of React Testing Library by Kent C. Dodds in such a way that: if the user does not interact with an element, it should not be tested.