test: Add reassurance in preexisting group components
Change multiple sections of the codebase to insure that no further bugs get pushed out. (1) Change the test script to no longer use "--silence" so that testers can read any errors that may be unexpected. This way the code should help reveal its own errors without arbitrarily stumbling upon them. Since this change, some tests then revealed underlying issues that are fixed such as: preventing forms from reloading the page, explicitly expressing and supressing that a test will throw an error in its testing block. (2) Add testing for when a group is being assigned a new name. The missing functionality allowed users to enter invalid blank names, and while in the mindset, the app now allows for padded whitespace as this causes little issue and gives the user more control over what to name their groups. (3) Finish adding tests for when a GroupTable is initially rendered to the screen to boost confidence in the preexisting code. (4) Finish organizing the blocks of code missed for organization in GroupTable's test suite. (5) Add tests for all components in the modal used for group splitting.