feat: Implement capability to remove members from group
Add a button for initiating the process for removing a member of a group. The process includes a disableable button when the groups has zero members. If the button is enabled and clicked, then a modal shows on screen and allows the user to select members that they wish to relinquish control of. Behind the scenes, this means that the member is removed as a reference of their former group and toggles the member's flag  to signify if they are grouped or not. This denotes that they are free to be picked up by another group. This whole feature allows for the user to remove a member that they no longer wish to be part of the assigned group.