refactor: Implement new properties for Group and Participant interfaces
Change the Group interface to use objects where possible instead of derived properties. This will stop all functions that use pursuers and distancers from having to use the 'id' property to find the correlating object -- cutting down on the work needed to retrieve data. More flexibility emerges from this since the object is usually what we care about when calling the former id properties. In cases where we don't care for the full object, we still have access to the id. Secondly, add properties to the Participant interface to assist in the tracking and calculation of actions during a chase. On this note, also, replace the factory methods with a builder object to allow for more flexible, unpredictable Group and Participant object construction.