feat: Implement controls to stop and start the chase
Let the user click on buttons that visually toggle to alert the user that a chase is started or stopped. Also, change the structure of TabbedDisplay so that it is more flexible and allows the content on the page to be more self-sufficient -- less is implied about what TabbedDisplay should be to the page as a whole such as a "main" element. Also, add the display of actions a participant has in a chase as well as the algorithm to help calculate such information. Each participant is assigned actions based on their chase and the participants that share that chase.