refactor: Update the ParticipantContainer and ParticipantTable files
Forced the ParticipantTable and ParticipantContainer files and their respective test suites to adhere to the testing guidelines. Made actionlisteners start with on- or handle- to help easily identify them and use the lint to enforce this rule. Also, properly regrouped related tests. Finally, simplified some props on ParticipantContainer so only one actionlistener is required for the component to communicate that the underlying data modal should be updated for the TabbedDisplay and consequently the entire app.