fix: Delay creating modals only when they are needed for display
Use the flags for halting modal display so that the DOM is not polluted with irrelevant elements. Also, fix a bug where refreshing the page caused a SSR error due to document not existing. The solution became using a class already provided by react-modal to the body when a modal is opened. The class is used to prevent scrolling while a modal is open.