style: Implement Material-UI
Commit to using Material-UI as a component library. The project follows the Material guidelines for displaying a component while also seeming to adhere to the WAI-ARIA guidelines. That conclusion in mind combined with the teams discovery of the Not-invented-here (NIH) syndrome has lead the project to adopt the Material-UI library to help focus on the goal of creating a usable and easy-to-understand companion app for Call of Cthulhu 7th Edition, instead of, making a WAI-ARIA compliant web app. This does not detract from that sentiment however. Accessibility is cool and allows for more people to enjoy the app and -- hopefully -- CoC7e. Also, some packages have been updated to support the newest version of Material-UI, such as Gatsby and its plugins.

Update Gatsby from 2.0 to 3.0.